feat: apply out-of-place FFT to enable better scheduling
FFT was previously implemented as an in-place op, requiring control-predecessors
to enforce execution order. Now that FFT is modeled as a pure, side-effect-free
operation in ZKIR, we apply it out-of-place and remove unnecessary control edges.

This change also reorders the `-tensor-ext-to-tensor` pass to run immediately
before `-one-shot-bufferize`, as required by:
https://github.com/zk-rabbit/zkir/pull/55